sửa lại backbone code theo sườn đã cho
